requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
networkx>=3.2
pydantic>=2.5.0
streamlit>=1.29.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0
crewai
psycopg2-binary
SQLAlchemy
pika
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
python-jose[cryptography]
passlib[bcrypt]
websockets
